DEV: Handle array of files or single file for uploadHandler (#21)
In core, we are changing the uploadHandler API to send through
multiple files at a time instead of just one. This commit allows
for either a single file or array until the core change is done.
We are selecting the first file in the array because that is how
it used to work with jQuery file uploader, which sent through one
file at a time, with the modal opening just one of the sent through
files.