FEATURE: Add support for downloading mp4 version
By default when uploading a video to mux regardless of format it is converted
to a streaming .m3u8 file. This change allows for mux to also store a
mp4 version of the upload file in addition to the streaming file. This
allows us to link directly to the mp4 video for downloading, but still
allow for streaming for optimal video playback.

This change includes two new site settings, one for enabling mp4
storage, and another one that adds a link to download the mp4 file below
the video player.