WIP: chat_message migration, update chat client on webhook event
This commit creates a new table for storing which chat messages have
videos.

Also this commit removes the js runloop checking if the video is ready.
Now the webhook will handle updating the client once it receives a
"ready" request. There is still some work to do here in that the client
is only updating if we actually edit the chat message. The webhook
currently doesn't modify the chat message, so to client isn't updating
at the moment.