FIX: Switch to unminified `hls.js` script (#38)
`hls.js` leans on `webworkify-webpack` to generate a Web Worker blob. A regex is used to parse the JS code for `require()` calls. For minified code, this results in some false-positives. One of those false-positives is named "init", which clashes with the name of a method which Ember adds to the Array prototype. The result is that some random Ember code is added to the `hls.js` Web Worker and causes a syntax error.

The unminified `hls.js` script does not suffer from the same false-positives, and so the issue is avoided.